requirements.txt
alembic==0.9.6
amqp==5.1.1
autopep8==1.7.0
Babel==2.5.1
billiard==4.2.0 #3.6.4.0
bio==1.3.8
biopython==1.79
biothings-client==0.2.6
blinker==1.4
boto3==1.21.44
botocore==1.24.44
celery== 5.4.0 #5.2.7 - updating celery b/c 5.2.7 has error with SQS udpates 
certifi==2017.7.27.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
dominate==2.3.1
elastic-transport==8.1.2
elasticsearch==8.1.2
email-validator==1.2.1
Flask==2.2.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
guess-language-spirit==0.5.3
idna==2.6
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
kombu==5.3.4
Mako==1.0.7
MarkupSafe==2.1.1
matplotlib==3.7.1
mygene==3.2.2
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.0.0
numpy==1.22.3
pandas==1.2.4
prompt-toolkit==3.0.29
protobuf==3.20.1
pycodestyle==2.9.1
pycurl==7.45.1
PyJWT==1.5.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.3
pytz==2022.1
redis==3.2.1
requests==2.27.1
rq==1.0
s3transfer==0.5.2
scipy==1.10.1
six==1.11.0
SQLAlchemy==1.3.18
supervisor==4.2.4
toml==0.10.2
tqdm==4.64.0
transaction==3.0.1
urllib3==1.26.9
uWSGI==2.0.20
vine==5.1.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.0
zipp==3.8.1
zope.interface==5.4.0
